+++ src/main/java/net/mohron/skyclaims/command/user/CommandReset.java
@@ -19,6 +19,7 @@
 package net.mohron.skyclaims.command.user;
 
 import net.mohron.skyclaims.command.CommandBase;
+import net.mohron.skyclaims.command.CommandIsland;
 import net.mohron.skyclaims.command.argument.Argument;
 import net.mohron.skyclaims.permissions.Options;
 import net.mohron.skyclaims.permissions.Permissions;
@@ -54,6 +55,7 @@ public class CommandReset extends CommandBase {
 
     public static void register() {
         try {
+            CommandIsland.addSubCommand(commandSpec, "reset");
             PLUGIN.getGame().getCommandManager().register(PLUGIN, commandSpec);
             PLUGIN.getLogger().debug("Registered command: CommandReset");
         } catch (UnsupportedOperationException e) {
@@ -88,9 +90,6 @@ public class CommandReset extends CommandBase {
         } else {
             player.getEnderChestInventory().clear();
             player.getInventory().clear();
-            if (PLUGIN.getConfig().getIntegrationConfig().getNucleus().isFirstJoinKit()) {
-                PLUGIN.getIntegration().getNucleus().ifPresent(n -> n.redeemFirstJoinKit(player));
-            }
 
             // Teleport any players located in the island's region to spawn
             Location<World> spawn = PLUGIN.getConfig().getWorldConfig().getWorld().getSpawnLocation();