+++ src/main/java/net/mohron/skyclaims/integration/nucleus/CommandHome.java
@@ -16,66 +16,65 @@
  * along with SkyClaims.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package net.mohron.skyclaims.command.user;
+package net.mohron.skyclaims.integration.nucleus;
 
+import io.github.nucleuspowered.nucleus.api.nucleusdata.Home;
+import io.github.nucleuspowered.nucleus.api.nucleusdata.NamedLocation;
 import net.mohron.skyclaims.command.CommandBase;
-import net.mohron.skyclaims.integration.Nucleus;
+import net.mohron.skyclaims.command.CommandIsland;
 import net.mohron.skyclaims.permissions.Permissions;
-import net.mohron.skyclaims.world.Island;
 import org.spongepowered.api.command.CommandException;
 import org.spongepowered.api.command.CommandResult;
 import org.spongepowered.api.command.CommandSource;
 import org.spongepowered.api.command.args.CommandContext;
 import org.spongepowered.api.command.spec.CommandSpec;
+import org.spongepowered.api.entity.Transform;
 import org.spongepowered.api.entity.living.player.Player;
+import org.spongepowered.api.entity.living.player.User;
 import org.spongepowered.api.text.Text;
 import org.spongepowered.api.text.format.TextColors;
+import org.spongepowered.api.world.World;
 
-public class CommandSetHome extends CommandBase {
+import java.util.Optional;
 
-    public static final String HELP_TEXT = "set your island home.";
+public class CommandHome extends CommandBase {
+
+    public static final String HELP_TEXT = "teleport to your home island.";
 
     public static CommandSpec commandSpec = CommandSpec.builder()
-        .permission(Permissions.COMMAND_SET_HOME)
+        .permission(Permissions.COMMAND_HOME)
         .description(Text.of(HELP_TEXT))
-        .executor(
-            (PLUGIN.getIntegration().getNucleus().isPresent() && PLUGIN.getConfig().getIntegrationConfig().getNucleus().isHomesEnabled())
-                ? new CommandSetHome()
-                : new CommandSetSpawn()
-        )
+        .executor(new CommandHome())
         .build();
 
     public static void register() {
         try {
+            CommandIsland.addSubCommand(commandSpec, "home");
             PLUGIN.getGame().getCommandManager().register(PLUGIN, commandSpec);
-            PLUGIN.getLogger().debug("Registered command: CommandSetHome");
+            PLUGIN.getLogger().debug("Registered command: CommandHome");
         } catch (UnsupportedOperationException e) {
             e.printStackTrace();
-            PLUGIN.getLogger().error("Failed to register command: CommandSetHome");
+            PLUGIN.getLogger().error("Failed to register command: CommandHome");
         }
     }
 
+    @Override
     public CommandResult execute(CommandSource src, CommandContext args) throws CommandException {
         if (!(src instanceof Player)) {
             throw new CommandException(Text.of("You must be a player to use this command!"));
         }
 
         Player player = (Player) src;
-        Island island = Island.get(player.getLocation())
-            .orElseThrow(() -> new CommandException(Text.of(TextColors.RED, "You must be on an island to set a home!")));
+        Transform<World> transform = getHome(player)
+            .orElseThrow(() -> new CommandException(Text.of(TextColors.RED, "You must set a home before using this command!")));
 
-        if (!island.hasPermissions(player)) {
-            throw new CommandException(Text.of(TextColors.RED, "You must have permission to set home on this island!"));
-        }
-
-        Nucleus nucleus = PLUGIN.getIntegration().getNucleus()
-            .orElseThrow(() -> new CommandException(Text.of(TextColors.RED, "Error: Home Command Requires Nucleus!")));
-
-        boolean success = nucleus.modifyOrCreateHome(player);
-        if (!success) {
-            throw new CommandException(Text.of(TextColors.RED, "An error was encountered while attempting to set your home!"));
-        }
+        player.setTransform(transform);
 
         return CommandResult.success();
+    }
+
+    private Optional<Transform<World>> getHome(User user) {
+        Optional<Home> oHome = NucleusIntegration.getHomeService().getHome(user, "Island");
+        return oHome.flatMap(NamedLocation::getTransform);
     }
 }