+++ CHANGELOG.md
@@ -1,9 +1,15 @@
 # Change Log
 
 ## Unreleased Features, Changes & Bugfixes
-**[Dev Builds](http://ci.devonthe.rocks/job/SkyClaims) - _Use at your own risk unless directed otherwise!_</br>**
-**REQUIRED: SF 2096+ & GP 292+ OPTIONAL: Nucleus 1.0.1+**
+**[Dev Builds](http://ci.devonthe.rocks/job/SkyClaims) - _Use at your own risk unless directed otherwise!_**
+**REQUIRED: SF build 2315+ & GP build 304+ OPTIONAL: Nucleus version 1.0.1+**
 - `/is reset` now automatically redeems the Nucleus First Join Kit, if present.
+
+## Beta 23
+**REQUIRED: SF build 2497+ & GP build 4.0.0.319+ OPTIONAL: Nucleus version 1.0.1+**
+- Updated GriefPrevention API. GP 4.0.0.319 is now the minimum required version.
+- Block-break/block-place flags in the wilderness of the SkyClaim's world are set to false automatically.
+- `/is expand` now cost 256 times more blocks due to GP's change to include y in claim block costs
 
 ## Beta 22
 **REQUIRED: SF 2096+ & GP 295+ OPTIONAL: Nucleus 1.0.1+**